app/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.7.0
boto3==1.34.41
botocore==1.34.41
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
Flask-Cors==4.0.0
ecdsa==0.18.0
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
pika==1.3.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2024.1
PyJWT==2.8.0
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.0.7
websockets==12.0
Werkzeug==3.0.1
SQLAlchemy-Utils==0.41.1
flask-swagger-ui==4.11.1
stripe==8.3.0